+++ store/cachekvstore.go
@@ -5,6 +5,7 @@ import (
 	"sort"
 	"sync"
 
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	cmn "github.com/tendermint/tmlibs/common"
 )
 
@@ -132,6 +133,16 @@ func (ci *cacheKVStore) Iterator(start, end []byte) Iterator {
 // Implements KVStore.
 func (ci *cacheKVStore) ReverseIterator(start, end []byte) Iterator {
 	return ci.iterator(start, end, false)
+}
+
+// Implements KVStore.
+func (ci *cacheKVStore) SubspaceIterator(prefix []byte) Iterator {
+	return ci.iterator(prefix, sdk.PrefixEndBytes(prefix), true)
+}
+
+// Implements KVStore.
+func (ci *cacheKVStore) ReverseSubspaceIterator(prefix []byte) Iterator {
+	return ci.iterator(prefix, sdk.PrefixEndBytes(prefix), false)
 }
 
 func (ci *cacheKVStore) iterator(start, end []byte, ascending bool) Iterator {