+++ store/dbstoreadapter.go
@@ -19,5 +19,13 @@ func (dsa dbStoreAdapter) CacheWrap() CacheWrap {
 	return NewCacheKVStore(dsa)
 }
 
+func (dsa dbStoreAdapter) SubspaceIterator(prefix []byte) Iterator {
+	return dsa.Iterator(prefix, sdk.PrefixEndBytes(prefix))
+}
+
+func (dsa dbStoreAdapter) ReverseSubspaceIterator(prefix []byte) Iterator {
+	return dsa.ReverseIterator(prefix, sdk.PrefixEndBytes(prefix))
+}
+
 // dbm.DB implements KVStore so we can CacheKVStore it.
 var _ KVStore = dbStoreAdapter{dbm.DB(nil)}