+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 0.15.0 (TBD)
+
+BREAKING CHANGES
+
+* Remove go-wire, use go-amino
+* [store] Add `SubspaceIterator` and `ReverseSubspaceIterator` to `KVStore` interface
+
+## 0.14.1 (April 9, 2018)
+
+BUG FIXES
+
+* [gaiacli] Fix all commands (just a duplicate of basecli for now)
+
 ## 0.14.0 (April 9, 2018)
 
 BREAKING CHANGES: